# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TOAN Store E-commerce API
  description: |
    API documentation for TOAN Store E-commerce platform.
    
    **Authentication**: JWT via HTTP-Only Cookies (`nike_auth_session` / `nike_admin_session`)  
    **Rate Limiting**: Redis-backed sliding window, Fail-Closed for sensitive endpoints.
  version: 2.5.0
  contact:
    email: admin@toanstore.com
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Get product detail
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List user's orders
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Orders]
      summary: Create new order
      security: [{ cookieAuth: [] }]
      responses:
        '201': { description: Order created }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nike_auth_session
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: nike_admin_session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        data: {}
        message: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...